Restructured the test packages.
