+++ backEnd/src/test/java/strand/StrandAnnotatorTest.java
@@ -1,4 +1,4 @@
-package genome;
+package strand;
 
 
 import org.junit.Before;
