+++ backEnd/src/test/java/strand/StrandTest.java
@@ -1,4 +1,4 @@
-package genome;
+package strand;
 
 import mutation.AbstractMutation;
 import strand.Strand;
